Refactor the site models out of main
Refactored the code to move the site models out of the main package
alone with the methods to populate the models. This change is to improve
the readability of the code base and prepare for upcoming rework.
