+++ .gitignore
@@ -1,6 +1,9 @@
 *.DS_Store
-assets
 build
 .vscode
 node_modules
 logs
+npm-debug.log
+
+# Generated site
+assets
